refactor(qbittorrent-e2e): extract workspace setup into workspace_setup module
